refactor(runtime): remove static import cycles
